Correct invalid entry in template
Previously, trying to declare the katello class on RedHat 7.3 node running
Puppet version 4.9.4 resulted in an error while evaluating the function call:
the reason was because the template was calling another function using the old
3.x.x style of `scope.function_versioncmp` instead of the newer style of
`scope.call_function`. This commit eliminates the need for calling the function
altogether by doing the conditional logic inside the manifest and passing
a variable, which is more "future proof" and easier to understand.